ps2emu-record: Record port numbers, use instead of IRQs for filtering
This records the recording of the port layout, and also adds recording
of port numbers. We can always count on the port number for the keyboard
being the same on any system, so this allows us to immediately know when
interrupts are coming from the keyboard or not, without having to use
the initial setup of the i8042 chip. In addition, it's unlikely we'll
actually need to replicate the port numbers exactly on the playback
machine, so since we don't even use it for IRQs anymore there's no sense
keeping it in here.