ps2emu-record: Remove the ability to record information from files
In order to get the information we need to properly record, we're going
to need to access stuff in procfs that we can't get at any other time
other then runtime, so trying to keep this feature in would be
impossible. Also it's not very useful now that the parsing portion
works.