Add basic support for version 1 of the logging format
This version of the logging format allows for the addition of more types of
information. Most importantly, the addition of sections. Sections allow us to
separate the main event recording from the initialization process, which
eventually can be used to play multiple series of events on the same device
without having to continually reinitialize the device.