ps2emu-replay: Sleep for half a second before replaying events
Before we seperated events and initialization, we'd have a slight delay
before playing any events due to the behavior of most PS/2 devices. When
we play the initialization sequence and then play the event sequence
directly after, it throws the device out of sync and usually results in
the driver getting confused and ignoring all the events coming from the
device. Sleeping for about half a second before we play the events
should fix this.