Remove the event origin from the log file format, fix keyboard replay
There's no point in recording the origin device in the log. This portion of the
log format was added before we had a complete understanding of how serio played
in with all of the other portions of the input system, and was never used in the
recorder. Here we get rid of it, and specify the type of device we're replaying
at the top of the file, allowing for actually working replay of keyboards in
addition to just touchpads and such.