ps2emu-record: Check for unmask_kbd_data
Version 4.3 of the Linux kernel added the default behavior of masking
all of the data coming out of the KBD port once a driver binds to it, so
that people's passwords aren't accidentally leaked into dmesg. In order
to record the keyboard properly, we need to check for the existence of
this feature in the kernel and turn it off while doing keyboard
recordings.