ps2emu-record.c: Don't output non-comments during normal recording
When we notify the user that the first stage of recording has completed
during a recording, we don't do so as a comment. While this isn't a
problem if we're forwarding to a text file, if we try to copy the output
of ps2emu-record from a terminal we run into the issue of ps2emu-replay
breaking as a result of ps2emu's notifications getting caught in the
log.