Fix compilation on Ubuntu 20.04
Not sure why, but Ubuntu 20.04 can not compile properly without this.

(see https://bugzilla.kernel.org/show_bug.cgi?id=209027#c5)

For the `ld` present in Ubuntu 20.04, the `-lglib-2.0` flags needs to be put at the end of the call, after the various `.o`.
While the current code works still fine on Fedora 32.

Anyway, with this, both Fedora 32 and Ubuntu 20.04 now compile.